FIX: Remove separate dependencies copy/cache steps
It seems the glob in these steps don't work for some versions of
docker if there is no file that satisfies them. As Google Cloud Run
isn't cacheing anything, the steps can be removed with no loss in
build performance in the Cloud (but there will be a loss if doing so
locally).